feat: promotion product list
